+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## Unreleased (develop)
 
-## 4.48.0 (staging)
+## 4.49.0 (staging)
+
+- added: Show swap KYC/terms modal for NExchange
+- added: Nym mixnet warning in Stake, Unstake, and Claim Rewards scenes
+- changed: Migrate Thorchain Savers and Thorchain Yield endpoints off NineRealms to gateway.liquify.com.
+
+## 4.48.0 (2026-04-23)
 
 - added: (iOS) Disable Liquid Glass UI for iOS 26 compatibility
 - added: Add dRPC and Subscan API key support